Added context-support, 1 part of changes suggested by zyga.
