+++ cmd/snap-confine/user-support.c
@@ -19,8 +19,6 @@
 
 #include <errno.h>
 #include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
 #include <sys/stat.h>
 
 #include "utils.h"
@@ -64,37 +62,4 @@ void setup_user_xdg_runtime_dir()
 	if (errno == 0 && chmod(xdg_runtime_dir, 0700) != 0) {
 		die("cannot change permissions of user XDG_RUNTIME_DIR directory to 0700");
 	}
-}
-
-void setup_snap_context_var()
-{
-	const char *snap_name = getenv("SNAP_NAME");
-	char context_path[1000];
-	// context is a 32 bytes, base64-encoding makes it 44.
-	char context_val[45];
-
-	if (snap_name == NULL) {
-		die("SNAP_NAME is not set");
-	}
-	// this should never happen.
-	if (strlen(CONTEXTS_DIR) + strlen(snap_name) + 2 > 1000) {
-		die("Cannot set SNAP_CONTEXT, file path too long");
-	}
-
-	strcpy(context_path, CONTEXTS_DIR "/");
-	strcat(context_path, snap_name);
-
-	FILE *f = fopen(context_path, "rt");
-	if (f == NULL) {
-		error
-		    ("Cannot open context file %s, SNAP_CONTEXT will not be set",
-		     context_path);
-		return;
-	}
-	if (fgets(context_val, 45, f) == NULL) {
-		error("Failed to read context file %s", context_path);
-	} else {
-		setenv("SNAP_CONTEXT", context_val, 1);
-	}
-	fclose(f);
 }
